15022022 - maze generator
added maze generator, node class, updated main.py, added requirements.txt for dependencies and project description headline in readme.
